dds_depth_sensor_proxy inherits from embedded_filter_sensor_interface
