+++ src/CMakeLists.txt
@@ -105,7 +105,6 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/labeled-points.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/to-string.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/eth-config-device.cpp"
-        "${CMAKE_CURRENT_LIST_DIR}/embedded-filter.cpp"
 
         "${CMAKE_CURRENT_LIST_DIR}/algo.h"
         "${CMAKE_CURRENT_LIST_DIR}/api.h"
