+++ src/dds/rs-dds-depth-sensor-proxy.h
@@ -4,6 +4,8 @@
 
 #include "rs-dds-sensor-proxy.h"
 #include <src/depth-sensor.h>
+#include <src/dds/rs-dds-depth-sensor-decimation-filter.h>
+#include <src/dds/rs-dds-depth-sensor-temporal-filter.h>
 
 
 namespace librealsense {
@@ -12,7 +14,7 @@ namespace librealsense {
 class dds_depth_sensor_proxy
     : public dds_sensor_proxy
     , public depth_stereo_sensor
-    , public embedded_decimation_filter_sensor
+    , public embedded_filter_sensor_interface
 {
     using super = dds_sensor_proxy;
 
@@ -28,12 +30,20 @@ public:
     float get_depth_scale() const override;
     float get_stereo_baseline_mm() const override;
 
+    // Override interface methods
+    void set(rs2_embedded_filter_type embedded_filter_type, std::vector<uint8_t> params) override;
+    std::vector<uint8_t> get(rs2_embedded_filter_type embedded_filter_type) override;
+    bool supports(rs2_embedded_filter_type embedded_filter_type) const override;
+
     bool extend_to( rs2_extension, void ** ptr ) override;  // extendable_interface
 
 protected:
     void add_no_metadata( frame *, streaming_impl & ) override;
     void add_frame_metadata( frame *, rsutils::json const & md, streaming_impl & ) override;
-};
 
+private:
+    std::shared_ptr<dds_depth_sensor_decimation_filter> _decimation_filter;
+    std::shared_ptr<dds_depth_sensor_temporal_filter> _temporal_filter;
+};
 
 }  // namespace librealsense
